backend/pyproject.toml
[build-system]
requires = ["setuptools>=75.6.0"]
build-backend = "setuptools.build_meta"

[project]
name = "naafi"
version = "0.1.0"
dependencies = [
    "Django==5.1.7",
    "django-configurations==2.5.1",
    "gunicorn==23.0.0",
    "newrelic==10.3.1",
    "psycopg2-binary==2.9.10",
    "dj-database-url==2.3.0",
    "django-model-utils==5.0.0",
    "django_unique_upload==0.2.1",
    "django-tenants==3.8.0",
    "djangorestframework==3.15.2",
    "django-filter==24.3",
    "drf-yasg==1.21.10",
    "django-storages==1.14.4",
    "boto3==1.35.76",
    "django-structlog==9.1.1",
    "structlog==25.4.0",
    "whitenoise>=6.9.0",
    "django-cors-headers>=4.7.0",
    "pillow>=11.2.1",
    "django-mptt>=0.17.0",
]

[project.optional-dependencies]
dev = [
    "ipdb==0.13.13",
    "ipython==8.30.0",
    "mkdocs==1.6.1",
    "flake8==7.1.1",
    "mock==5.1.0",
    "factory-boy==3.3.1",
    "pytest-django==4.9.0",
    "coverage==7.6.9",
    "django-stubs>=3.33.0",
    "mypy>=1.5.1"
]

[tool.ruff]
# Ruff configuration
select = ["E", "F", "W", "I"]  # Enable pycodestyle (E, W), PyFlakes (F), and isort (I) rules
ignore = []  # No rules to ignore
line-length = 88  # Match black's default line length
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    "__pycache__",
    "migrations",
]

[tool.ruff.isort]
known-first-party = ["naafi"]  # Replace with your project name
lines-between-types = 1
combine-as-imports = true

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true
follow_imports = "silent"
strict_optional = true